Drop the blanket lint suppressions and rewrite the index loops
`rspice-core` carried a 22-lint crate-scope `#![allow(...)]` added to reach a
green CI run, and 55 more module-scope suppressions sat on top of it across
core, matrix and conformance. A suppression added to make a gate pass hides
every future instance of the same defect, so all 56 are gone.

This change clears the mechanical classes the removal exposed:
`needless_range_loop` becomes an iterator, `collapsible_if` a let-chain,
`never_loop` a `slice.first()` guard (those loops always returned on their
first iteration), plus `manual_strip`, `manual_contains`, `derivable_impls`,
`unnecessary_cast`, `needless_lifetimes` and `excessive_precision`. Every
rewrite is order- and value-preserving: the float literals clippy shortened
parse to the same `f64`, and no loop changed the sequence it visits.

The three `approx_constant` sites keep their literals under a function-scope
allow: 1.57079 is XSPICE's published `angle` ceiling and 0.6366 a
Kirschning-Jansen dispersion coefficient, not roundings of a std constant.

+++ crates/rspice-core/src/circuit/linear_stamping.rs
@@ -898,8 +898,8 @@ impl CircuitData {
             node
         }
         let mut union = |positive: NodeId, negative: NodeId| {
-            let positive = find(&mut parent, positive as usize);
-            let negative = find(&mut parent, negative as usize);
+            let positive = find(&mut parent, positive);
+            let negative = find(&mut parent, negative);
             if positive != negative {
                 parent[positive] = negative;
             }
@@ -936,8 +936,8 @@ impl CircuitData {
                 seed_current[index] = true;
                 continue;
             }
-            let positive_root = find(&mut parent, positive as usize);
-            let negative_root = find(&mut parent, negative as usize);
+            let positive_root = find(&mut parent, positive);
+            let negative_root = find(&mut parent, negative);
             if positive_root == negative_root {
                 seed_current[index] = true;
                 has_inductor_cycle = true;
